requirements.txt
numpy<2
chromadb==1.5.4
torch==2.2.2
transformers==4.44.2
tokenizers==0.19.1
safetensors==0.7.0
huggingface_hub==0.36.2
google-cloud-aiplatform==1.140.0
google-genai==1.66.0
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
pyyaml
regex==2026.2.28
networkx==3.6.1
tenacity==9.1.4
